LayoutTests/fast/filesystem/workers/simple-temporary-sync.html

<html>
<head>
<script src="../resources/fs-worker-common.js"></script>
<script src="../resources/fs-worker-test-pre.js"></script>
</head>
<body>
<p id="description">This tests requesting TEMPORARY FileSystemSync in Worker.</p>
<div id="console"></div>
<script>
startWorker('../resources/simple-temporary-sync.js');
</script>
<script src="../resources/fs-worker-test-post.js"></script>
</body>
</html>
