LayoutTests/fast/filesystem/workers/simple-temporary-sync-expected.txt

This tests requesting TEMPORARY FileSystemSync in Worker.

Starting worker: ../resources/simple-temporary-sync.js
Description: requestFileSystemSync TEMPORARY test.
Successfully obtained TEMPORARY FileSystem:file__0:Temporary
PASS: fileSystem.name.length > 0 is true
PASS: fileSystem.root.fullPath is "/"
DONE

TEST COMPLETE
